+++ lib/sink/connection/client_connection_handler.ex
@@ -9,6 +9,7 @@ defmodule Sink.Connection.ClientConnectionHandler do
   @type offset() :: non_neg_integer()
   @type event_data() :: binary()
   @type message_id() :: non_neg_integer()
+  @type nack_data() :: {binary(), Strint.t()}
 
   @doc """
   The connection has been established and authenticated
@@ -24,6 +25,11 @@ defmodule Sink.Connection.ClientConnectionHandler do
   Run implementer's logic for handling a "ack"
   """
   @callback handle_ack(ack_key()) :: :ok
+
+  @doc """
+  Run implementer's logic for handling a "nack"
+  """
+  @callback handle_nack(ack_key(), nack_data()) :: :ok
 
   @doc """
   Run implementer's logic for handling a "publish" message