+++ lib/sink/connection/protocol.ex
@@ -4,13 +4,6 @@ defmodule Sink.Connection.Protocol do
     <<message_type_id::4, message_id::integer-size(12)>>
   end
 
-  def encode_frame(:publish, message_id, payload) do
-    message_type_id = 4
-    header = <<message_type_id::4, message_id::integer-size(12)>>
-
-    header <> payload
-  end
-
   def encode_frame(:ping) do
     message_type_id = 5
     message_id = 0
@@ -23,6 +16,18 @@ defmodule Sink.Connection.Protocol do
     <<message_type_id::4, message_id::integer-size(12)>>
   end
 
+  def encode_frame(message_type, message_id, payload) do
+    message_type_id =
+      case message_type do
+        :publish -> 4
+        :nack -> 7
+      end
+
+    header = <<message_type_id::4, message_id::integer-size(12)>>
+
+    header <> payload
+  end
+
   def decode_frame(message) do
     <<message_type_id::4, message_id::integer-size(12), payload::binary>> = message
 
@@ -31,6 +36,7 @@ defmodule Sink.Connection.Protocol do
       4 -> {:publish, message_id, payload}
       5 -> :ping
       6 -> :pong
+      7 -> {:nack, message_id, payload}
     end
   end
 
@@ -43,8 +49,20 @@ defmodule Sink.Connection.Protocol do
       event_data
   end
 
-  def encode_payload(:ack, message_type_id, message_id) do
-    <<message_type_id::4, message_id::integer-size(12)>>
+  def encode_payload(:nack, {machine_message, human_message}) do
+    Varint.LEB128.encode(byte_size(machine_message)) <>
+      machine_message <>
+      human_message
+  end
+
+  def decode_payload(:ack, <<message_type_id::4, message_id::integer-size(12)>>) do
+    {message_type_id, message_id}
+  end
+
+  def decode_payload(:nack, <<payload::binary>>) do
+    {machine_message, human_message} = _parse_varint_delimited_value(payload)
+
+    {machine_message, human_message}
   end
 
   def decode_payload(:publish, payload) do